pyproject.toml
[project]
name = "homey-assistant-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[deepgram,elevenlabs,google,mcp,openai,silero]~=1.0",
    "livekit-plugins-noise-cancellation~=0.2",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
]
